Library

Documenting the public user interface.

TurbulentWords.jl

TurbulentWords.TurbulentWords — Module

Create turbulent animations from words using Oceananigans.jl

TurbulentWords.alternating — Method

alternating(N)

Return an array of alternating 1 and -1 of length N.

Example

julia> using TurbulentWords

julia> alternating(5)
5-element Vector{Int64}:
  1
 -1
  1
 -1
  1

TurbulentWords.letter_to_array — Function

letter_to_array(letter, Ny=biggest_Ny; hpad=20)

Convert letter to an array of ones and zeros.

TurbulentWords.word_to_array — Method

word_to_array(word;
              greek = false,
              multiplicative_factors = ones(length(word)),
              word_Ny = size(letter_to_array(word[1]), 2),
              letter_pad::Int = 20,
              vpad::Int = 0,
              pad_to_square::Bool = false,
              hpad::Int = 50)

Convert word to an array of ones and zeros.

TurbulentWords.word_to_simulation — Method

word_to_simulation(word;
                   dynamics = :two_dimensional_turbulence,
                   other_kw...)

Return a simulations with word as initial condition.